anaconda/requirements.txt
adapters==0.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.2
blis==0.7.11
bs4==0.0.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
dnspython==2.6.1
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.4
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.25.2
orjson==3.10.3
packaging==24.0
pandas==2.1.0
pillow==10.3.0
preshed==3.0.9
psycopg==3.1.18
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytube==15.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
thinc==8.2.4
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.1
transformers==4.39.3
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.3
ujson==5.10.0
urllib3==2.0.4
uvicorn==0.30.1
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
